==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Gives step by step instructions on how to access and download the 
# lung cancer dataset from the OSF dataverse database. 
# Author: Lexi Knight
# Date: 28 March 2024
# Contact: Lexi.Knight@mail.utoronto.ca
# License: MIT


### Steps to Download data ###

# 1. Go to https://osf.io/search?activeFilters=%5B%5D&sort=-relevance&view_only=
# 2. Type in search bar "Lung Cancer"
# 3. Select the project "Lung cancer Predictions" by Sicong Chen(Magic)
# 4. Under files, click  plus sign on "OSF Storage (Australia - Sydney) to expand files 
# 5. Click "Lung_Cancer_Meta.xlsx" 
# 6. There are two tabs containing two differnt datasets, select "clinical_info" 
# and click download the dataset as an excel file.
